+++ template.go
@@ -1,11 +1,11 @@
 package model
 
 type Template struct {
-	TemplateID        int64  `json:"template_id" bson:"template_id"`
+	TemplateID        int    `json:"template_id" bson:"template_id"`
 	ImageURL          string `json:"image_url" bson:"image_url"`
 	BackgroundURL     string `json:"background_url" bson:"background_url"`
 	BackgroundColor   string `json:"background_color" bson:"background_color"`
-	PrintfileID       int64  `json:"printfile_id" bson:"printfile_id"`
+	PrintfileID       int    `json:"printfile_id" bson:"printfile_id"`
 	TemplateWidth     int    `json:"template_width" bson:"template_width"`
 	TemplateHeight    int    `json:"template_height" bson:"template_height"`
 	PrintAreaWidth    int    `json:"print_area_width" bson:"print_area_width"`
